requirements.txt
fastapi==0.115.12
mangum==0.19.0
pdf2image==1.17.0
boto3==1.37.26
botocore==1.37.26
urllib3==1.26.18  # Compatible with botocore 1.37.26
python-multipart==0.0.20
python-dotenv==1.1.0